AcademicRiskPredictor/pyproject.toml
[project]
name = "academicriskpredictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.0",
    "joblib>=1.5.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "shap>=0.49.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]
